moved all remaining stuff from Slang::Math to LIBC::math and LIBC::stdlib
