+++ Projects/Extensions/LIBC/stdlib/stdlib.cpp
@@ -33,7 +33,9 @@ void stdlib_t::initialize( ExtensionNamespace* scope )
 
 void stdlib_t::provideMethods( ExtensionMethods& methods )
 {
-    methods.push_back( new ABS() );
+    methods.push_back( new ABSDouble() );
+    methods.push_back( new ABSFloat() );
+    methods.push_back( new ABSInt() );
     methods.push_back( new RAND() );
     methods.push_back( new SRAND() );
 }
