--- Projects/Extensions/System/Math/Abs.h
@@ -1,113 +1,0 @@
-
-#ifndef Slang_Extensions_System_Math_Abs_h
-#define Slang_Extensions_System_Math_Abs_h
-
-
-// Library includes
-#include <cmath>
-
-// Project includes
-#include <Core/Common/Exceptions.h>
-#include <Core/Designtime/BuildInTypes/DoubleObject.h>
-#include <Core/Designtime/BuildInTypes/FloatObject.h>
-#include <Core/Designtime/BuildInTypes/IntegerObject.h>
-#include <Core/Extensions/ExtensionMethod.h>
-#include <Core/Runtime/BuildInTypes/DoubleObject.h>
-#include <Core/Runtime/BuildInTypes/FloatObject.h>
-#include <Core/Runtime/BuildInTypes/IntegerObject.h>
-#include <Core/Runtime/BuildInTypes/StringObject.h>
-#include <Core/Tools.h>
-#include <Core/VirtualMachine/Controller.h>
-
-// Forward declarations
-
-// Namespace declarations
-
-
-namespace Slang {
-namespace Extensions {
-namespace System {
-namespace Math {
-
-
-class AbsDouble: public ExtensionMethod
-{
-public:
-	AbsDouble()
-	: ExtensionMethod(0, "abs", Designtime::DoubleObject::TYPENAME)
-	{
-		ParameterList params;
-		params.push_back(Parameter::CreateDesigntime("value", Common::TypeDeclaration(Designtime::DoubleObject::TYPENAME)));
-
-		setSignature(params);
-	}
-
-public:
-	Runtime::ControlFlow::E execute(Common::ThreadId threadId, const ParameterList& params, Runtime::Object* result, const Token& token)
-	{
-		ParameterList list = mergeParameters(params);
-
-		try {
-			ParameterList::const_iterator it = list.begin();
-
-			double param_value = (*it++).value().toDouble();
-
-			*result = Runtime::DoubleObject(std::abs(param_value));
-		}
-		catch ( std::exception& e ) {
-			Runtime::Object *data = Controller::Instance().repository()->createInstance(Runtime::StringObject::TYPENAME);
-			*data = Runtime::StringObject(std::string(e.what()));
-
-			Controller::Instance().thread(threadId)->exception(data, token.position());
-			return Runtime::ControlFlow::Throw;
-		}
-
-		return Runtime::ControlFlow::Normal;
-	}
-};
-
-
-class AbsFloat: public ExtensionMethod
-{
-public:
-	AbsFloat()
-	: ExtensionMethod(0, "abs", Designtime::FloatObject::TYPENAME)
-	{
-		ParameterList params;
-		params.push_back(Parameter::CreateDesigntime("value", Common::TypeDeclaration(Designtime::FloatObject::TYPENAME)));
-
-		setSignature(params);
-	}
-
-public:
-	Runtime::ControlFlow::E execute(Common::ThreadId threadId, const ParameterList& params, Runtime::Object* result, const Token& token)
-	{
-		ParameterList list = mergeParameters(params);
-
-		try {
-			ParameterList::const_iterator it = list.begin();
-
-			float param_value = (*it++).value().toFloat();
-
-			*result = Runtime::FloatObject(std::abs(param_value));
-		}
-		catch ( std::exception& e ) {
-			Runtime::Object *data = Controller::Instance().repository()->createInstance(Runtime::StringObject::TYPENAME);
-			*data = Runtime::StringObject(std::string(e.what()));
-
-			Controller::Instance().thread(threadId)->exception() = Runtime::ExceptionData(data, token.position());
-			return Runtime::ControlFlow::Throw;
-		}
-
-		return Runtime::ControlFlow::Normal;
-	}
-};
-
-
-}
-}
-}
-}
-
-
-#endif
