--- Projects/Extensions/System/Math/CMakeLists.txt
@@ -1,23 +1,0 @@
-INCLUDE(${PROJECT_SOURCE_DIR}/CMake/Base.cmake)
-
-SET(SOURCES
-	MathExtension.cpp
-)
-
-SET(HEADERS
-	MathExtension.h
-)
-
-SET(DEPENDENCIES
-	Utils
-	Core
-)
-
-SET(MODULES
-)
-
-add_include(Projects/Core)
-add_include(Projects/Extensions)
-add_include(Projects/Utils)
-
-build_static_lib(ExtensionsSystemMath "${MODULES}")
