--- Projects/Extensions/System/Math/Sqrt.h
@@ -1,108 +1,0 @@
-
-#ifndef Slang_Extensions_System_Math_Sqrt_h
-#define Slang_Extensions_System_Math_Sqrt_h
-
-
-// Library includes
-#include <math.h>
-
-// Project includes
-#include <Core/Designtime/BuildInTypes/DoubleObject.h>
-#include <Core/Designtime/BuildInTypes/FloatObject.h>
-#include <Core/Extensions/ExtensionMethod.h>
-#include <Core/Runtime/BuildInTypes/DoubleObject.h>
-#include <Core/Runtime/BuildInTypes/FloatObject.h>
-#include <Core/Tools.h>
-
-// Forward declarations
-
-// Namespace declarations
-
-
-namespace Slang {
-namespace Extensions {
-namespace System {
-namespace Math {
-
-
-class SqrtDouble: public ExtensionMethod
-{
-public:
-	SqrtDouble()
-	: ExtensionMethod(0, "sqrt", Designtime::DoubleObject::TYPENAME)
-	{
-		ParameterList params;
-		params.push_back(Parameter::CreateDesigntime("value", Designtime::DoubleObject::TYPENAME));
-
-		setSignature(params);
-	}
-
-public:
-	Runtime::ControlFlow::E execute(Common::ThreadId threadId, const ParameterList& params, Runtime::Object* result, const Token& token)
-	{
-		ParameterList list = mergeParameters(params);
-
-		try {
-			ParameterList::const_iterator it = list.begin();
-
-			double param_value = (*it++).value().toDouble();
-
-			*result = Runtime::DoubleObject(sqrt(param_value));
-		}
-		catch ( std::exception& e ) {
-			Runtime::Object *data = Controller::Instance().repository()->createInstance(Runtime::StringObject::TYPENAME, ANONYMOUS_OBJECT);
-			*data = Runtime::StringObject(std::string(e.what()));
-
-			Controller::Instance().thread(threadId)->exception() = Runtime::ExceptionData(data, token.position());
-			return Runtime::ControlFlow::Throw;
-		}
-
-		return Runtime::ControlFlow::Normal;
-	}
-};
-
-
-class SqrtFloat: public ExtensionMethod
-{
-public:
-	SqrtFloat()
-	: ExtensionMethod(0, "sqrt", Designtime::FloatObject::TYPENAME)
-	{
-		ParameterList params;
-		params.push_back(Parameter::CreateDesigntime("value", Designtime::FloatObject::TYPENAME));
-
-		setSignature(params);
-	}
-
-public:
-	Runtime::ControlFlow::E execute(Common::ThreadId threadId, const ParameterList& params, Runtime::Object* result, const Token& token)
-	{
-		ParameterList list = mergeParameters(params);
-
-		try {
-			ParameterList::const_iterator it = list.begin();
-
-			double param_value = (*it++).value().toDouble();
-
-			*result = Runtime::DoubleObject(sqrt(param_value));
-		}
-		catch ( std::exception& e ) {
-			Runtime::Object *data = Controller::Instance().repository()->createInstance(Runtime::StringObject::TYPENAME, ANONYMOUS_OBJECT);
-			*data = Runtime::StringObject(std::string(e.what()));
-
-			Controller::Instance().thread(threadId)->exception() = Runtime::ExceptionData(data, token.position());
-			return Runtime::ControlFlow::Throw;
-		}
-
-		return Runtime::ControlFlow::Normal;
-	}
-};
-
-
-}
-}
-}
-}
-
-
-#endif
